harness.sh: fix call stack printing on bash < 4.4
Before bash-4.4 functions imported from the environment did not
appear in BASH_SOURCE / FUNCNAME / BASH_LINENO (and probably more).
As such, hard coding that the last two levels of calls were to be
hidden (they are supposed to be calls to bunny's run_suite() and
run_test()) was a mistake. It meant that two levels of the call stack
were missing when ran with bash-4.3 and lower.

Signed-off-by: ypsah <asyph@tutanota.com>